ChemicalElements

This is an extremely simple module used as the base brick of AuChimiste. Below we go through the whole of its functionalities in just a few lines of code. In general you only need to worry about using ChemicalElements directly if your calculations require isotopes to be added. The default table of elements provides access only to stable elements.

Database manipulation

You can get a list of available atomic symbols with list_elements. Suppose you want to check if deuterium D is present in the list, you can use its symbol for inspection:

:D ∈ list_elements()

false

Since it is not present but your calculations require this isotope, you feed the database with add_element; you also decide to add tritium. In fact add_element will not fail if the element exists, but issue a warning. You can try adding an existing element to see what happens:

add_element("D", "deuterium", 1, 2.0141017781)
add_element("Tr", "tritium", 1, 3.0160492820)
add_element("H", "hydrogen", 1, 1.008)

┌ Warning: The provided element H is already present in the elements dictionary. If you are trying to create an isotope, please chose a different name. Also notice that default stable elements cannot be modified.
└ @ ChemicalElements ~/work/AuChimiste.jl/AuChimiste.jl/src/ChemicalElements.jl:220

If you wish to get back to the standard data you can do so with reset_elements_table. Notice below that deuterium mass is no longer available.

reset_elements_table()
has_element(:D)

false

Element data retrieval

It is possible to retrieve the atomic_mass. Other data retrieval functions include atomic_number and element_name. All of these work with both string and symbols.

atomic_mass(:C), atomic_number(:C), element_name(:C)

(12.011, 6, "carbon")

Getting the whole element data can be achieved at once as follows:

element(:Cl)

Cl (17, chlorine) 35.45 kg/kmol

In this case is also possible to query the data through the atomic number:

element(26)

Fe (26, iron) 55.845 kg/kmol

On the other hand atomic masses of unstable elements are not accessible:

try atomic_mass(:Po) catch e; @error(e) end

┌ Error: NoIsotopeProvidedError: Accessing the atomic mass of unstable element Po is not supported. Please consider creating a named isotope of this element with `add_isotope`.
└ @ Main ChemicalElements.md:59

To have an unstable element listed, you need to add_isotope before. For instance, let's add Po-187 to the database.

add_isotope(:Po, 187.003030)

Po187 (84, polonium-187) 187.00303 kg/kmol

API

ChemicalElements.has_element — Function

has_element(e::Union{String, Symbol})

Check if element exists in list of atomic symbols.

ChemicalElements.list_elements — Function

list_elements()

Provides access to the list of atomic symbols.

ChemicalElements.reset_elements_table — Function

reset_elements_table()

Remove any user-defined element.

ChemicalElements.add_element — Function

add_element(
    symbol::String,
    name::String,
    number::Int64,
    mass::Float64;
    verbose = true
)

Create chemical element name with associated symbol and atomic number. The value of atomic mass is given in grams per mole.

ChemicalElements.add_isotope — Function

add_isotope(
    symbol::String,
    mass::Float64;
    name = nothing,
    verbose = true
)

Create isotope of element symbol with provided mass in grams per mole. If isothope is known by a specific name then use it instead of a name-mass naming scheme.

ChemicalElements.atomic_mass — Function

atomic_mass(e::Union{String,Symbol})

Atomic mass of element [g/mol].

ChemicalElements.atomic_number — Function

atomic_number(e::Union{String,Symbol})

Atomic number of element.

ChemicalElements.element_name — Function

element_name(e::Union{String,Symbol})

Element name from atomic symbol.

ChemicalElements.element — Function

element(e::Int64)
element(e::Union{String,Symbol})

Element data from symbol or number.