Private API

This part of the documentation is intended for developers. It might also be useful for standard users trying to understand bugs or propose features. AuChimiste aims at having 100% first-level entities documented so that design features can be understood in the future.

Development rules

Code written, code documented, code tested.

Code lines make 72 characters, never more than 79.

Code is not cluttered and comments are minimal.

Code abuses of multiple dispatch if needed.

Code is Julia, nothing else.

Chemical Elements

AuChimiste.ELEMENTS — Constant

Default table of elements. This table should not be modified by any internal or external operation. Although it is declared as constant, that means simply that ELEMENTS cannot be attributed to, but the resulting dictionary may be accidentally modified.

AuChimiste.USER_ELEMENTS — Constant

Runtime modifiable table of elements. All operations must be performed in this table so that user-defined elements (isothopes) can be made available. This is the table to be internally modified and read by all functions requiring to access data.

AuChimiste.handle_element — Function

handle_element(f, e)

Applies function f to element e. This function wraps the call of f with a standardized error-handling used accross the module.

AuChimiste.find_element — Function

find_element(v, prop)

Find element for which property prop has value v.

Chemical Components

Chemical Kinetics

Chemical Reactors

Combustion Chemistry

Physical Chemistry

AuChimiste.mean_molecular_mass_y — Function

mean_molecular_mass_y(Y, W)

Mean molecular mass computed from mass fractions.

AuChimiste.mean_molecular_mass_x — Function

mean_molecular_mass_x(X, W)

Mean molecular mass computed from mole fractions.

Chemical Thermodynamics

Exception types

AuChimiste.NoSuchElementError — Type

Element (or isotope) was not found in user database.

AuChimiste.NoIsotopeProvidedError — Type

Unstable elements do not provide atomic mass.

AuChimiste.EmptyCompositionError — Type

A composition set is missing for the given component.

AuChimiste.InvalidScalerError — Type

The provided scaler targets an unspecified element.