Users

Chemical Elements

AuChimiste.has_element — Function

has_element(e::Union{String, Symbol})

Check if element exists in list of atomic symbols.

AuChimiste.list_elements — Function

list_elements()

Provides access to the list of atomic symbols.

AuChimiste.reset_elements_table — Function

reset_elements_table()

Remove any user-defined element.

AuChimiste.add_element — Function

add_element(
    symbol::String,
    name::String,
    number::Int64,
    mass::Float64;
    verbose = true
)

Create chemical element name with associated symbol and atomic number. The value of atomic mass is given in grams per mole.

AuChimiste.add_isotope — Function

add_isotope(
    symbol::String,
    mass::Float64;
    name = nothing,
    verbose = true
)

Create isotope of element symbol with provided mass in grams per mole. If isothope is known by a specific name then use it instead of a name-mass naming scheme.

AuChimiste.atomic_mass — Function

atomic_mass(e::AtomicData)
atomic_mass(e::Union{String,Symbol})

Atomic mass of element [g/mol].

AuChimiste.atomic_number — Function

atomic_number(e::AtomicData)
atomic_number(e::Union{String,Symbol})

Atomic number of element.

AuChimiste.element_name — Function

element_name(e::AtomicData)
element_name(e::Union{String,Symbol})

Element name from atomic symbol.

AuChimiste.element — Function

element(e::Int64)
element(e::Union{String,Symbol})

Element data from symbol or number.

AuChimiste.AtomicData — Type

Represents a chemical element.

Fields

symbol::String: Element symbol in periodic table.

name::String: Element name in periodic table.

number::Int64: Element number in atomic units.

mass::Float64: Element atomic mass [kg/kmol].

Chemical Components

AuChimiste.ChemicalComponent — Type

Represents a chemical component.

Fields

elements::Vector{Symbol}: Array of component symbols.

coefficients::Vector{Float64}: Array of stoichiometric coefficients.

mole_fractions::Vector{Float64}: Array of elemental mole fractions.

mass_fractions::Vector{Float64}: Array of elemental mass fractions.

molar_mass::Float64: Molar mass of corresponding stoichiometry.

charge::Number: Global charge of component.

Notes

This structure is not intended to be called as a constructor, safe use of its features require using component construction in combination with a composition specification.

The array of elements is unsorted when construction is performed through component but may get rearranged when composing new chemical components through supported algebra.

Care must be taken when using molar_mass because it is given for the associated coefficients. That is always the expected behavior for molecular components but might not be the case in other applications (solids, solutions) when the mean molecular mass may be required.

AuChimiste.ComponentQuantity — Type

Represents a quantity of component.

Fields

mass::Float64: Mass of component in arbitrary units.

composition::AuChimiste.ChemicalComponent: Elemental composition of component.

AuChimiste.component — Function

component(spec; kw...)
component(c::Composition{Stoichiometry}, charge)
component(c::Composition{MoleProportion}, charge)
component(c::Composition{MassProportion}, charge)

Compile component from given composition specification. This function is a wrapper eliminating the need of calling stoichiometry, mole_proportions or mass_proportions directly. The value of spec must be the symbol representing one of their names.

AuChimiste.stoichiometry — Function

stoichiometry(; kw...)

Create composition based on elemental stoichiometry.

AuChimiste.mole_proportions — Function

mole_proportions(; scale = nothing, kw...)

Create composition based on relative molar proportions. The main different w.r.t. stoichiometry is the presence of a scaling factor to correct stoichiometry representation of the given composition.

AuChimiste.mass_proportions — Function

mass_proportions(; scale = nothing, kw...)

Create composition based on relative molar proportions. This is essentially the same thing as mole_proportions but in this case the element keywords are interpreted as being the mass proportions ofa associated elements.

AuChimiste.stoichiometry_map — Function

stoichiometry_map(c::ChemicalComponent)

Returns component map of elemental stoichiometry.

AuChimiste.mole_fractions_map — Function

mole_fractions_map(c::ChemicalComponent)

Returns component map of elemental mole fractions.

AuChimiste.mass_fractions_map — Function

mass_fractions_map(c::ChemicalComponent)

Returns component map of elemental mass fractions.

AuChimiste.quantity — Function

quantity(c::ChemicalComponent, mass::Float64)
quantity(spec::Symbol, mass::Float64; kw...)

Creates a quantity of chemical component. It may be explicit, i.e. by providing directly a ChemicalComponent, or implicit, that means, by creating a component directly from its chemical composition and specification method (wrapping component).

The following are not exported but is worth the end-user to known them:

AuChimiste.CompositionTypes — Type

Provides specification of allowed chemical composition types, which are used to declare compositions in terms of one of the following specification methods:

Stoichiometry: stoichiometric coefficients

MoleProportion: molar proportions of elements

MassProportion: mass proportions of elements

AuChimiste.Composition — Type

Creates a typed composition specification for later construction of chemical component with component. Generally the end-user is not expected to use this structure directly, wrappers being provided by the available composition types through functions stoichiometry, mole_proportions, and mass_proportions.

Fields

data::NamedTuple: Tuple of elements and their amounts.

scale::Pair{Symbol, <:Number}: Scaler element and coefficient for construction of component.

Chemical Kinetics

Chemical Reactors

Combustion Chemistry

Physical Chemistry

AuChimiste.mean_molecular_mass_y — Function

mean_molecular_mass_y(Y, W)

Mean molecular mass computed from mass fractions.

AuChimiste.mean_molecular_mass_x — Function

mean_molecular_mass_x(X, W)

Mean molecular mass computed from mole fractions.

AuChimiste.mean_molecular_mass — Function

mean_molecular_mass(U, W; basis)

Compute mean molecular mass based on given composition data.

AuChimiste.get_mole_fractions — Function

get_mole_fractions(Y, W)
get_mole_fractions(Y, W, M)

Get mole fractions from mass fractions.

AuChimiste.get_mass_fractions — Function

get_mass_fractions(X, W)
get_mass_fractions(X, W, M)

Get mass fractions from mole fractions.

Chemical Thermodynamics

Database parsing

Constants

AuChimiste.ELECTRON_MASS — Constant

Electron Mass $m_e$ [9.109382915e-31 kg]

AuChimiste.AVOGADRO — Constant

Avogadro's Number $N_\mathrm{A}$ [6.02214076e26 number/kmol]

AuChimiste.GAS_CONSTANT — Constant

Ideal gas constant $R$ [8.31446261815324 J/(mol.K)].

AuChimiste.STEFAN_BOLTZMANN — Constant

Stefan-Boltzmann constant $\sigma$ [5.670374419e-8 W/(m².K⁴)].

AuChimiste.P_NORMAL — Constant

Reference atmospheric pressure [101325.0 Pa].

AuChimiste.T_NORMAL — Constant

Normal atmospheric temperature [273.15 K].

AuChimiste.T_STANDARD — Constant

Standard atmospheric temperature [298.15 K].

Exception types

Exception types are not exported as they are not intended for other purposes than handling errors inside AuChimiste. The documentation below is provided so that the end-user can better understand their occurrence for debugging their own code or reporting bugs:

AuChimiste.NoSuchElementError — Type

Element (or isotope) was not found in user database.

AuChimiste.NoIsotopeProvidedError — Type

Unstable elements do not provide atomic mass.

AuChimiste.EmptyCompositionError — Type

A composition set is missing for the given component.

AuChimiste.InvalidScalerError — Type

The provided scaler targets an unspecified element.