Public API

Chemical Elements

AuChimiste.has_element — Function

has_element(e::Union{String, Symbol})

Check if element exists in list of atomic symbols.

AuChimiste.list_elements — Function

list_elements()

Provides access to the list of atomic symbols.

AuChimiste.reset_elements_table — Function

reset_elements_table()

Remove any user-defined element.

AuChimiste.add_element — Function

add_element(
    symbol::String,
    name::String,
    number::Int64,
    mass::Float64;
    verbose = true
)

Create chemical element name with associated symbol and atomic number. The value of atomic mass is given in grams per mole.

AuChimiste.add_isotope — Function

add_isotope(
    symbol::String,
    mass::Float64;
    name = nothing,
    verbose = true
)

Create isotope of element symbol with provided mass in grams per mole. If isothope is known by a specific name then use it instead of a name-mass naming scheme.

AuChimiste.atomic_mass — Function

atomic_mass(e::AtomicData)
atomic_mass(e::Union{String,Symbol})

Atomic mass of element [g/mol].

AuChimiste.atomic_number — Function

atomic_number(e::AtomicData)
atomic_number(e::Union{String,Symbol})

Atomic number of element.

AuChimiste.element_name — Function

element_name(e::AtomicData)
element_name(e::Union{String,Symbol})

Element name from atomic symbol.

AuChimiste.element — Function

element(e::Int64)
element(e::Union{String,Symbol})

Element data from symbol or number.

AuChimiste.AtomicData — Type

Represents a chemical element.

Fields

symbol::String: Element symbol in periodic table.

name::String: Element name in periodic table.

number::Int64: Element number in atomic units.

mass::Float64: Element atomic mass [kg/kmol].

Chemical Components

AuChimiste.stoichiometry — Function

stoichiometry(; kw...)

Create composition based on elemental stoichiometry.

AuChimiste.mole_proportions — Function

mole_proportions(; scale = nothing, kw...)

Create composition based on relative molar proportions.

AuChimiste.mass_proportions — Function

mole_proportions(; scale = nothing, kw...)

Create composition based on relative molar proportions.

AuChimiste.component — Function

component(spec; kw...)

Compile component from given composition specification.

Chemical Kinetics

Chemical Reactors

Combustion Chemistry

Physical Chemistry

AuChimiste.mean_molecular_mass — Function

mean_molecular_mass(U, W; basis)

Compute mean molecular mass based on given composition data.

AuChimiste.get_mole_fractions — Function

get_mole_fractions(Y, W)
get_mole_fractions(Y, W, M)

Get mole fractions from mass fractions.

AuChimiste.get_mass_fractions — Function

get_mass_fractions(X, W)
get_mass_fractions(X, W, M)

Get mass fractions from mole fractions.

Chemical Thermodynamics